Replace babel-cli with babel-require
Should allow for faster builds
