+++ package.json
@@ -4,7 +4,7 @@
   "description": "partyq",
   "main": "index.js",
   "scripts": {
-    "start": "babel-node index.js",
+    "start": "node index.js",
     "test:server": "mocha --compilers js:babel/register --require ./server/test/test_helper.js --recursive 'server/test/**/*.@(js|jsx)'",
     "test:client": "mocha --compilers js:babel-core/register --require ./app/test/test_helper.js --recursive 'app/test/**/*.@(js|jsx)'",
     "test": "npm run test:common && npm run test:client",
@@ -46,13 +46,13 @@
   "devDependencies": {
     "autoprefixer": "6.0.3",
     "babel": "6.1.18",
-    "babel-cli": "6.2.0",
     "babel-core": "6.2.1",
     "babel-eslint": "4.1.5",
     "babel-loader": "6.2.0",
     "babel-preset-es2015": "6.1.18",
     "babel-preset-react": "6.1.18",
     "babel-preset-stage-2": "6.1.18",
+    "babel-register": "6.2.0",
     "chai": "3.4.0",
     "css-loader": "0.21.0",
     "eslint": "1.7.3",
